Refactor rate limiter implementation and update imports to use the new paths
